+++ web/src/defaults.browser.ts
@@ -6,10 +6,13 @@ import {
 } from '@/platform/BrowserConfigStorage';
 import { ContentStorage } from '@/platform/types/ContentStorage';
 import { BrowserContentStorage } from '@/platform/BrowserContentStorage';
+import type { DefaultTokenStorageProps } from '@/platform/types/DefaultTokenStorageProps';
 
 /** Default token storage for browser environments. */
-export function defaultTokenStorage(pid: string, tag?: string): TokenStorage {
-  return new BrowserTokenStorage(pid, tag);
+export function defaultTokenStorage(
+  props: DefaultTokenStorageProps,
+): TokenStorage {
+  return new BrowserTokenStorage(props.pid, props.tag);
 }
 
 /** Default content storage for browser environments. */