+++ web/src/defaults.ts
@@ -3,10 +3,13 @@ import { NodeTokenStorage } from '@/platform/NodeTokenStorage';
 import { readConfigNode, saveConfigNode } from '@/platform/NodeConfigStorage';
 import { ContentStorage } from '@/platform/types/ContentStorage';
 import { NodeContentStorage } from '@/platform/NodeContentStorage';
+import type { DefaultTokenStorageProps } from '@/platform/types/DefaultTokenStorageProps';
 
 /** Default token storage for Node.js environments. */
-export function defaultTokenStorage(pid: string, tag?: string): TokenStorage {
-  return new NodeTokenStorage(pid, tag);
+export function defaultTokenStorage(
+  props: DefaultTokenStorageProps,
+): TokenStorage {
+  return new NodeTokenStorage(props.pid, props.tag);
 }
 
 /** Default content storage for Node.js environments. */