+++ web/src/network/websocket/BeamWebSocket.ts
@@ -10,7 +10,7 @@ import {
 } from '@/__generated__/apis';
 import { HttpRequester } from '@/network/http/types/HttpRequester';
 
-interface BeamWebSocketConnect {
+interface BeamWebSocketConnectParams {
   requester: HttpRequester;
   cid: string;
   pid: string;
@@ -18,12 +18,12 @@ interface BeamWebSocketConnect {
 }
 
 export class BeamWebSocket {
+  private url = '';
+  private cid = '';
+  private pid = '';
+  private refreshToken = '';
   private socket?: WebSocket;
   private requester?: HttpRequester;
-  private url?: string;
-  private cid?: string;
-  private pid?: string;
-  private refreshToken?: string;
   private connectPromiseWithResolvers?: PromiseWithResolversPolyfill;
   private isDisconnecting = false;
   private isReconnecting = false;
@@ -59,43 +59,47 @@ export class BeamWebSocket {
     this.socket = socket;
 
     // Web socket open event handler
-    socket.onopen = () => {
-      this.reconnectAttempts = 0;
-      this.connectPromiseWithResolvers?.resolve();
-    };
-
+    socket.onopen = () => this.handleOpen();
     // Web socket error event handler
-    socket.onerror = async (e) => {
-      console.error('WebSocket error:', e);
-      if (
-        socket.readyState === WebSocket.OPEN ||
-        socket.readyState === WebSocket.CONNECTING
-      ) {
-        // If the socket is still open or connecting, we can try to reconnect
-        socket.close();
-      } else {
-        this.connectPromiseWithResolvers?.reject(
-          new BeamWebSocketError('WebSocket error occurred'),
-        );
-      }
-    };
-
+    socket.onerror = (event) => this.handleError(event);
     // Web socket close event handler
-    this.socket.onclose = async (e) => {
-      // if explicitly called disconnect(), don't reconnect
-      if (this.isDisconnecting) return;
+    socket.onclose = (event) => this.handleClose(event);
+  }
 
-      console.warn('WebSocket closed:', e.code, e.reason);
-      if (this.reconnectAttempts < this.maxRetries) {
-        await this.reconnect();
-      } else {
-        this.connectPromiseWithResolvers?.reject(
-          new BeamWebSocketError(
-            'Maximum web socket reconnect attempts reached',
-          ),
-        );
-      }
-    };
+  private handleOpen() {
+    this.reconnectAttempts = 0;
+    this.connectPromiseWithResolvers?.resolve();
+  }
+
+  private handleError(e: Event) {
+    if (
+      this.socket?.readyState === WebSocket.OPEN ||
+      this.socket?.readyState === WebSocket.CONNECTING
+    ) {
+      // If the socket is still open or connecting, we can try to reconnect
+      this.socket.close();
+    } else {
+      this.connectPromiseWithResolvers?.reject(
+        new BeamWebSocketError('WebSocket error occurred', { cause: e }),
+      );
+    }
+  }
+
+  private async handleClose(e: CloseEvent) {
+    // if explicitly called disconnect(), don't reconnect
+    if (this.isDisconnecting) return;
+
+    console.warn('WebSocket closed:', e.code, e.reason);
+    if (this.reconnectAttempts < this.maxRetries) {
+      await this.reconnect();
+    } else {
+      this.connectPromiseWithResolvers?.reject(
+        new BeamWebSocketError(
+          'Maximum web socket reconnect attempts reached',
+          { cause: e },
+        ),
+      );
+    }
   }
 
   private async setWebSocketUrl(): Promise<void> {
@@ -146,14 +150,14 @@ export class BeamWebSocket {
   /**
    * Opens a WebSocket connection to the Beamable server.
    *
-   * @param {BeamWebSocketConnect} param - The connection parameters.
+   * @param {BeamWebSocketConnectParams} params - The connection parameters.
    * @returns {Promise<void>}
    */
-  async connect(param: BeamWebSocketConnect): Promise<void> {
-    this.requester = param.requester;
-    this.cid = param.cid;
-    this.pid = param.pid;
-    this.refreshToken = param.refreshToken;
+  async connect(params: BeamWebSocketConnectParams): Promise<void> {
+    this.requester = params.requester;
+    this.cid = params.cid;
+    this.pid = params.pid;
+    this.refreshToken = params.refreshToken;
     this.isDisconnecting = false;
     this.reconnectAttempts = 0;
     this.connectPromiseWithResolvers = promiseWithResolvers();
@@ -163,6 +167,7 @@ export class BeamWebSocket {
 
   private async reconnect(): Promise<void> {
     if (this.isReconnecting) return;
+
     this.isReconnecting = true;
     this.reconnectAttempts++;
     const jitter = Math.random() * 500; // up to 0.5 seconds of randomness
@@ -170,6 +175,7 @@ export class BeamWebSocket {
     await wait(delay); // pause before the next attempt
 
     if (this.isDisconnecting) return;
+
     await this.initWebSocket();
     this.isReconnecting = false;
   }